requirements/common.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.9.1
pydantic-settings==2.4.0
aiokafka==0.10.0
protobuf==4.25.3
grpcio==1.62.1
structlog==24.4.0
python-json-logger==2.0.7
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation-sqlalchemy==0.48b0
prometheus-client==0.20.0
